+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
-## 1.0.0 (22-Jul-2022)
+## 0.8.0 (01-Sept-2023)
+- Created a confirm dialog class
+- Confirm before quitting